# projects/notification-management-api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing email, SMS, and push notifications",
        "title": "Notification Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/email/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List email notifications with optional date and status filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "List emails",
                "parameters": [
                    {
                        "enum": [
                            "PENDING",
                            "SENT"
                        ],
                        "type": "string",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 50,
                        "minimum": 0,
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339)",
                        "name": "create_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/email.GetAllEmailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error_handling.Errors"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Queue a single email notification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Send email",
                "parameters": [
                    {
                        "description": "Send Email Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/email.SendEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error_handling.Errors"
                        }
                    }
                }
            }
        },
        "/api/v1/email/batch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Queue multiple email notifications (max 1000)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Send batch emails",
                "parameters": [
                    {
                        "description": "Send Batch Email Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/email.SendBatchEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error_handling.Errors"
                        }
                    }
                }
            }
        },
        "/api/v1/email/cancel": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cancel email notifications matching the given status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Cancel emails",
                "parameters": [
                    {
                        "enum": [
                            "PENDING"
                        ],
                        "type": "string",
                        "description": "Status to cancel",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error_handling.Errors"
                        }
                    }
                }
            }
        },
        "/api/v1/email/status": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the status of a single email notification",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Get email status by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Email ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/email.GetEmailStatusByIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error_handling.Errors"
                        }
                    }
                }
            }
        },
        "/api/v1/email/status/batch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the status of multiple email notifications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Get email statuses by IDs",
                "parameters": [
                    {
                        "description": "Batch Status Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/email.GetStatusByBatchIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/email.GetStatusByBatchIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error_handling.Errors"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "email.Email": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "integer"
                },
                "from": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "idempotency_key": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "integer"
                },
                "sent_at": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "email.GetAllEmailResponse": {
            "type": "object",
            "required": [
                "page",
                "page_size"
            ],
            "properties": {
                "emails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/email.Email"
                    }
                },
                "page": {
                    "type": "integer",
                    "minimum": 1
                },
                "page_size": {
                    "type": "integer",
                    "maximum": 50,
                    "minimum": 0
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "email.GetEmailStatusByIDResponse": {
            "type": "object",
            "properties": {
                "email_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "email.GetStatusByBatchIDRequest": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "email.GetStatusByBatchIDResponse": {
            "type": "object",
            "properties": {
                "emails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/email.GetEmailStatusByIDResponse"
                    }
                }
            }
        },
        "email.SendBatchEmail": {
            "type": "object",
            "required": [
                "content",
                "from",
                "priority",
                "subject",
                "to",
                "type"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 10000,
                    "minLength": 1
                },
                "from": {
                    "type": "string",
                    "maxLength": 255
                },
                "priority": {
                    "type": "string",
                    "enum": [
                        "LOW",
                        "MEDIUM",
                        "HIGH"
                    ]
                },
                "scheduled_at": {
                    "type": "string"
                },
                "subject": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                },
                "to": {
                    "type": "string",
                    "maxLength": 255
                },
                "type": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                }
            }
        },
        "email.SendBatchEmailRequest": {
            "type": "object",
            "properties": {
                "emails": {
                    "type": "array",
                    "maxItems": 1000,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/email.SendBatchEmail"
                    }
                }
            }
        },
        "email.SendEmailRequest": {
            "type": "object",
            "required": [
                "content",
                "from",
                "priority",
                "subject",
                "to",
                "type"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 10000,
                    "minLength": 1
                },
                "from": {
                    "type": "string",
                    "maxLength": 255
                },
                "priority": {
                    "type": "string",
                    "enum": [
                        "LOW",
                        "MEDIUM",
                        "HIGH"
                    ]
                },
                "scheduled_at": {
                    "type": "string"
                },
                "subject": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                },
                "to": {
                    "type": "string",
                    "maxLength": 255
                },
                "type": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                }
            }
        },
        "error_handling.Errors": {
            "type": "object",
            "properties": {
                "CorrelationID": {
                    "type": "string"
                },
                "Err": {},
                "ErrorType": {
                    "type": "string"
                },
                "Message": {
                    "type": "string"
                },
                "Parameter": {
                    "type": "string"
                },
                "StackTrace": {
                    "type": "string"
                },
                "StatusCode": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}